move /v2/searchextended over to tapir
