+++ webapi/src/main/scala/org/knora/webapi/routing/v2/SearchRouteV2.scala
@@ -12,6 +12,7 @@ import org.apache.pekko.http.scaladsl.server.RouteResult
 import zio.*
 
 import scala.concurrent.Future
+
 import dsp.errors.BadRequestException
 import dsp.valueobjects.Iri
 import org.knora.webapi.*
