+++ webapi/src/main/scala/org/knora/webapi/slice/common/api/ApiV2.scala
@@ -1,0 +1,50 @@
+package org.knora.webapi.slice.common.api
+
+import org.knora.webapi.ApiV2Complex
+import org.knora.webapi.ApiV2Schema
+
+object ApiV2 {
+
+  /**
+   * The names of the supported Http Headers.
+   */
+  object Headers {
+
+    /**
+     * The name of the HTTP header that can be used to specify how markup should be returned with
+     * text values.
+     */
+    val xKnoraAcceptMarkup: String = "x-knora-accept-markup"
+
+    /**
+     * The name of the HTTP header in which results from a project can be requested.
+     */
+    val xKnoraAcceptProject: String = "x-knora-accept-project"
+
+    /**
+     * The name of the HTTP header in which an ontology schema can be requested.
+     */
+    val xKnoraAcceptSchemaHeader: String = "x-knora-accept-schema"
+
+    /**
+     * The name of the HTTP header that can be used to request hierarchical or flat JSON-LD.
+     */
+    val xKnoraJsonLdRendering: String = "x-knora-json-ld-rendering"
+  }
+
+  object QueryParams {
+
+    /**
+     * The name of the URL parameter in which an ontology schema can be requested.
+     */
+    val schema: String = "schema"
+
+    /**
+     * The name of the URL parameter that can be used to specify how markup should be returned
+     * with text values.
+     */
+    val markup: String = "markup"
+  }
+
+  val defaultApiV2Schema: ApiV2Schema = ApiV2Complex
+}
