+++ webapi/src/main/scala/org/knora/webapi/slice/common/api/KnoraResponseRenderer.scala
@@ -1,0 +1,30 @@
+package org.knora.webapi.slice.common.api
+
+import sttp.model.MediaType
+import zio.Task
+import zio.ZIO
+import zio.ZLayer
+
+import org.knora.webapi.ApiV2Schema
+import org.knora.webapi.Rendering
+import org.knora.webapi.SchemaRendering
+import org.knora.webapi.config.AppConfig
+import org.knora.webapi.messages.util.rdf.RdfFormat
+import org.knora.webapi.messages.v2.responder.KnoraResponseV2
+import org.knora.webapi.slice.common.api.KnoraResponseRenderer.FormatOptions
+import org.knora.webapi.slice.common.api.KnoraResponseRenderer.RenderedResponse
+
+final class KnoraResponseRenderer(config: AppConfig) {
+  def render(response: KnoraResponseV2, opts: FormatOptions): Task[(RenderedResponse, MediaType)] =
+    ZIO.attempt(response.format(opts, config)).map((_, opts.format.mediaType))
+}
+
+object KnoraResponseRenderer {
+  type RenderedResponse = String
+  final case class FormatOptions(format: RdfFormat, schema: ApiV2Schema, rendering: Set[Rendering])
+  object FormatOptions {
+    def from(f: RdfFormat, s: SchemaRendering): FormatOptions = FormatOptions(f, s.schema, s.rendering)
+  }
+
+  val layer = ZLayer.derive[KnoraResponseRenderer]
+}
